refactor: extract seprateScript
